+++ RELEASES.md
@@ -6,19 +6,21 @@ This release focus on improving and extending data samplers.
 
 ### New features
 
-- SingleShotMemorySampler is available
+- `SingleShotMemorySampler()` is available
+
 - Data samplers now support:
   - The use of a data augmenter
   - Having a warmup period before augmentation
 
 ### Major improvement
 
+- `single_lookup()` now returns the rank of each match.
+- Memory Samplers now scale to millions of examples without slowdown.
 - Rewrote the sampler interface to allows better customization and subclassing.
 
 ### Major bug fixes
 
 - Fixed the issue where model couldn't be reloaded due to a bug in load_model()
-
 
 ## 0.6 - Distance Metrics